Create TitanBulkLoader.java
A way to add multi-properties to Titan using ScriptInputFormat - this file lets you split the name of an input property on the colon character ":" and if the first half is the word "set", map the property to Titan with set cardinality. If it is "list", map the property to Titan with "list" cardinality. If there is no colon character in the property key, then just call regular getOrCreateVertexProperty - note that this is basically a high-jacking of the code in this repo - https://github.com/dkuppitz/openflights

Further note that this is obviously a stupid way to do this, and something like this should be done that just accounts for how the property was set in the script being fed to ScriptInputFormat, and I really can't even recall if this was "necessary" but I think it was to get Titan's multi-properties working...